Move `BitmapCache` class to ui package
